refactor: expect one template for every version
Now one template is defined for every version, following a pattern of
plibv4-<distribution><version> instead of having one template for every
distribution with versions defined as sidecars.

While this will be more tedious to maintain, the alternative is that every
sidecar (=version definition) is started on each run for one distro+version
combination which is a terrible waste.

I'm still pondering on a more elegant structure, but so far, all options suck
to some extent.